// generated/types/order.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getAddressDecoder,
  getAddressEncoder,
  getI64Decoder,
  getI64Encoder,
  getStructDecoder,
  getStructEncoder,
  getU32Decoder,
  getU32Encoder,
  getU64Decoder,
  getU64Encoder,
  type Address,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";
import {
  getOrderSideDecoder,
  getOrderSideEncoder,
  getTokenTypeDecoder,
  getTokenTypeEncoder,
  type OrderSide,
  type OrderSideArgs,
  type TokenType,
  type TokenTypeArgs,
} from ".";

export type Order = {
  id: bigint;
  marketId: number;
  userKey: Address;
  side: OrderSide;
  tokenType: TokenType;
  price: bigint;
  quantity: bigint;
  filledquantity: bigint;
  timestamp: bigint;
};

export type OrderArgs = {
  id: number | bigint;
  marketId: number;
  userKey: Address;
  side: OrderSideArgs;
  tokenType: TokenTypeArgs;
  price: number | bigint;
  quantity: number | bigint;
  filledquantity: number | bigint;
  timestamp: number | bigint;
};

export function getOrderEncoder(): FixedSizeEncoder<OrderArgs> {
  return getStructEncoder([
    ["id", getU64Encoder()],
    ["marketId", getU32Encoder()],
    ["userKey", getAddressEncoder()],
    ["side", getOrderSideEncoder()],
    ["tokenType", getTokenTypeEncoder()],
    ["price", getU64Encoder()],
    ["quantity", getU64Encoder()],
    ["filledquantity", getU64Encoder()],
    ["timestamp", getI64Encoder()],
  ]);
}

export function getOrderDecoder(): FixedSizeDecoder<Order> {
  return getStructDecoder([
    ["id", getU64Decoder()],
    ["marketId", getU32Decoder()],
    ["userKey", getAddressDecoder()],
    ["side", getOrderSideDecoder()],
    ["tokenType", getTokenTypeDecoder()],
    ["price", getU64Decoder()],
    ["quantity", getU64Decoder()],
    ["filledquantity", getU64Decoder()],
    ["timestamp", getI64Decoder()],
  ]);
}

export function getOrderCodec(): FixedSizeCodec<OrderArgs, Order> {
  return combineCodec(getOrderEncoder(), getOrderDecoder());
}